+++ src/main/java/com/gregtechceu/gtceu/integration/ae2/machine/MEOutputHatchPartMachine.java
@@ -45,7 +45,7 @@ public class MEOutputHatchPartMachine extends MEHatchPartMachine implements IInW
 
     @Override
     @NotNull
-    protected NotifiableFluidTank createTank(long initialCapacity, int slots, Object... args) {
+    protected NotifiableFluidTank createTank(int initialCapacity, int slots, Object... args) {
         this.internalBuffer = new SerializableGenericStackInv(this::onChanged, slots);
         return new InaccessibleInfiniteSlot(this, this.internalBuffer);
     }
